Case study: The Fremont Bridge Bicycle Counter
The Fremont Bridge Bicycle Counter records the number of bikes that cross the bridge using the pedestrian/bicycle pathways. Inductive loops on the east and west pathways count the passing of bicycles regardless of travel direction. 
The data consists of: 
•	a date/time field: Date
•	east pathway count field: Fremont Bridge NB 
•	west pathway count field: Fremont Bridge SB 
The count fields represent the total bicycles detected during the specified one hour period. Direction of travel is not specified, but in general most traffic in the Fremont Bridge NB field is travelling northbound and most traffic in the Fremont Bridge SB field is travelling southbound.